pyproject.toml
[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyannote-openvino"
version = "0.1.0"
description = "Drop-in, OpenVINO-accelerated speaker diarization for pyannote.audio."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "numpy>=2.2.6",
    "scipy>=1.17.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "pyannote-audio>=4.0.4",
    "openvino>=2026.0.0",
    "optimum>=2.1.0",
    "optimum-intel>=1.27.0",
    "optimum-onnx>=0.1.0",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.1",
]

[project.optional-dependencies]
stt = [
    "openai-whisper>=20250625",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pyannote_openvino"]
